+++ src/commands/remove_directory.rs
@@ -1,4 +1,7 @@
+use crate::types::command_response;
+use serde_json::Value;
 use std::{fs, io, path::Path};
+use warp;
 
 pub fn remove_directory(path: &Path, force: Option<bool>) -> Result<(), io::Error> {
     if path.is_dir() {
@@ -22,4 +25,49 @@ pub fn remove_directory(path: &Path, force: Option<bool>) -> Result<(), io::Erro
     }
 
     Ok(())
+}
+
+pub fn interface(parameters: &Option<Value>) -> Option<command_response::Response> {
+    if let Some(params) = parameters {
+        let force = params.get("force").and_then(|v| v.as_bool()).unwrap_or(false);
+
+        if let Some(path_str) = params.get("path").and_then(Value::as_str) {
+            let path = Path::new(path_str);
+
+            match remove_directory(path, Some(force)) {
+                Ok(()) => {
+                    return Some(command_response::Response {
+                        data: None,
+                        status: warp::http::StatusCode::OK.into(),
+                        success: true,
+                        message: "Directory was successfully removed".to_string(),
+                        error: None,
+                    })
+                }
+                Err(error) => {
+                    return Some(command_response::Response {
+                        data: None,
+                        status: warp::http::StatusCode::INTERNAL_SERVER_ERROR.into(),
+                        success: false,
+                        message: "Failed to remove directory".to_string(),
+                        error: Some(command_response::Error {
+                            r#type: Some(command_response::ErrorType::InvalidRequest),
+                            message: format!("Error removing directory: {}", error).to_string(),
+                        }),
+                    })
+                }
+            }
+        }
+    }
+
+    Some(command_response::Response {
+        data: None,
+        status: warp::http::StatusCode::BAD_REQUEST.into(),
+        success: false,
+        message: "Missing parameters".to_string(),
+        error: Some(command_response::Error {
+            r#type: Some(command_response::ErrorType::InvalidRequest),
+            message: "Required parameters are missing".to_string(),
+        }),
+    })
 }